Update authors in pyproject.toml and README.md, enhance logging in quickrun.py, and introduce amber_files.py for AMBER file utilities. Refactor run_mmpbsa.py to streamline MMPBSA input file creation and validation processes. Improve output directory structure in README.md for clarity
